Adding some handlers to Controller.
Added `hello_in`, `send_features_request`, `features_reply_in` event
handlers to the Controller, in order to handle handshake messages/events
prior to Switch instantiation. This was done because the switch will be
created after receiving FeaturesReply message, that will be handled on
features_reply_in handler. As it needs access to the controller, we
decided to keep it on the controller instead of an external App.