Adds a new event (KycoRawMessageOutError)
KycoMessageOutError inherits from KycoAppEvent, so it can not be added
into the raw buffer. As it can be generated on the tcp_handler (when
trying to read/recv on the connection), and the tcp_handler can only
access the raw buffer, so I've created the "KycoRawMessageOutError" to
work as a wrapper for KycoMessageOutError.