Fix metadata access on setup.py
FIX #316

Previously we were importing the metadata.py module as a python module.
But, by doing so we also call the kytos/core/__init__.py file, that
tries to import kytos.core.controller module and, as a consequence,
lot's of dependencies (such as Flask), that hasn't yet being installed.
So, this is why we need to parse the 'metadata.py' file as a text file.
I've also renamed from _metadata.py to metadata.py to make it cleaner.