Calling the NApp setup during the NApp init.
Since the `setup()` method works like the `__init__()`, the NApp
developer set some variables to be used on the NApp, expecting them to
be available as soon as the NApp (object) is ready. So, it is much
better to run this setup on the end of the `__init__` method instead of
the `run`.